Organizando um pouco das páginas
