--- src/resources/screens/ProfileScreen.js
@@ -1,17 +1,0 @@
-import { View, Text, StyleSheet } from "react-native";
-
-export default function ProfileScreen() {
-  return (
-    <View style={styles.container}>
-      <Text>Profile</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-});
