+++ src/resources/screens/categories_screen/CategoriesScreen.js
@@ -1,0 +1,10 @@
+import { View, Text } from "react-native";
+import { styles } from "./CategoriesScreenStyle";
+
+export default function CategoriesScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>Departaments</Text>
+    </View>
+  );
+}
