+++ src/resources/screens/categories_screen/CategoriesScreenStyle.js
