+++ src/resources/screens/new_categories_screen/NewCategory.js
