+++ src/resources/screens/new_categories_screen/NewCategoryStyle.js
