+++ src/resources/screens/product_details_screen/ProductDetails.js
@@ -1,0 +1,9 @@
+import { View, Text, StyleSheet } from "react-native";
+
+export default function ProductDetailsScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>Detalhes do produto</Text>
+    </View>
+  );
+}
