+++ src/resources/screens/product_list/ProductList.js
@@ -1,0 +1,9 @@
+import { View, Text, StyleSheet } from "react-native";
+
+export default function ProductListScreen() {
+  return (
+    <View style={styles.container}>
+      <Text>Lista de produtos</Text>
+    </View>
+  );
+}
