+++ src/resources/screens/profile_screen/ProfileScreenStyle.js
@@ -1,0 +1,21 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  vwButtons: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  titulo: {
+    fontSize: 30,
+    marginBottom: 40,
+  },
+  btnNav: {
+    borderBlockColor: "rgb(255, 125, 3)",
+    borderWidth: 4,
+    borderRadius: 0,
+    padding: 15,
+  },
+  
+});
